fix discrete levels overflow (#53)
Counting the number of discrete levels for types where that was greater than `usize::max_value()` caused an overflow, when the only reason this was ever implemented was just for histogram equalisation to figure out if the output for a pixel should be rounded or not. Moved to `is_integral` for that and removed `discrete_levels`. Will also do a release as well